Create StaticInputParts.tsx and move static components from TextInput.tsx into there; Move action icon buttons components from TextInput.tsx to InputActionButtons.tsx with refactoring; Fix imports and usages of refactored components
