+++ client/src/components/Input/InputRedux.tsx
@@ -4,7 +4,11 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faSpinner } from '@fortawesome/free-solid-svg-icons';
 
 import { ErrorBadge, COLOR_TYPE } from './ErrorBadge';
-import { ClearInput, CopyButton, ShowPassword } from './InputActionButtons';
+import {
+  ClearButton,
+  CopyButton,
+  ShowPasswordIcon,
+} from './InputActionButtons';
 
 import classes from './Input.module.scss';
 
@@ -79,12 +83,7 @@ const InputRedux: React.FC<Props> = props => {
   });
 
   const onClearInputClick = () => {
-    onClose && onClose(false);
     onChange('');
-  };
-
-  const onShowPassClick = () => {
-    toggleShowPass(!showPass);
   };
 
   const hasError =
@@ -121,29 +120,31 @@ const InputRedux: React.FC<Props> = props => {
           type={showPass ? 'text' : type}
           data-clear={clearInput || showCopyButton}
         />
-        {(clearInput || onClose) && !loading && (
-          <ClearInput onClick={onClearInputClick} type={type} uiType={uiType} />
-        )}
-        {clearInput && type === 'password' && (
-          <ShowPassword
-            showPass={showPass}
-            onClick={onShowPassClick}
-            uiType={uiType}
-          />
-        )}
-        {showCopyButton && (
-          <CopyButton
-            onClick={async () => {
-              try {
-                const data = await navigator.clipboard.readText();
-                onChange(data);
-              } catch (e) {
-                console.error('Paste error: ', e);
-              }
-            }}
-            uiType={uiType}
-          />
-        )}
+        <ClearButton
+          isVisible={(clearInput || onClose) && !loading}
+          onClear={onClearInputClick}
+          onClose={onClose}
+          inputType={type}
+          uiType={uiType}
+        />
+        <ShowPasswordIcon
+          isVisible={clearInput && type === 'password'}
+          showPass={showPass}
+          toggleShowPass={toggleShowPass}
+          uiType={uiType}
+        />
+        <CopyButton
+          isVisible={showCopyButton}
+          onClick={async () => {
+            try {
+              const data = await navigator.clipboard.readText();
+              onChange(data);
+            } catch (e) {
+              console.error('Paste error: ', e);
+            }
+          }}
+          uiType={uiType}
+        />
         {loading && (
           <FontAwesomeIcon
             icon={faSpinner}
